Move value validation to their own package
