+++ app/classes/Validation/Rules/Value/NotNull.php
@@ -1,5 +1,7 @@
 <?php
-namespace KCMS\Validation\Rules;
+namespace KCMS\Validation\Rules\Value;
+
+use KCMS\Validation\Rules\AbstractRule;
 
 class NotNull extends AbstractRule
 {
